LayoutTests/fast/loader/location-port-expected.txt

This tests that modifying location.port works as it should

PASS: internalFrame.contentWindow.location.port == "" should be true and is.
PASS: internalFrame.contentWindow.location.port == "88" should be true and is.
PASS: internalFrame.contentWindow.location.port == "" should be true and is.
PASS: internalFrame.contentWindow.location.port == "80" should be true and is.
